[major] break down quantum_state.py into a module of several files
